feat: доработка комментариев

+++ ExploreWithMeServer/src/main/java/ru/practicum/explorewithme/services/impl/CommentServiceImpl.java
@@ -2,6 +2,8 @@ package ru.practicum.explorewithme.services.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.explorewithme.controllers.exceptionHandling.exceptions.ConditionsNotMetException;
 import ru.practicum.explorewithme.controllers.exceptionHandling.exceptions.EntryNotFoundException;
 import ru.practicum.explorewithme.dto.comment.CommentDto;
 import ru.practicum.explorewithme.dto.comment.mapper.CommentMapper;
@@ -9,34 +11,29 @@ import ru.practicum.explorewithme.model.Comment;
 import ru.practicum.explorewithme.model.Event;
 import ru.practicum.explorewithme.model.User;
 import ru.practicum.explorewithme.repositories.CommentRepository;
+import ru.practicum.explorewithme.repositories.EventRepository;
+import ru.practicum.explorewithme.repositories.UserRepository;
 import ru.practicum.explorewithme.services.CommentService;
-import ru.practicum.explorewithme.services.EventService;
-import ru.practicum.explorewithme.services.UserService;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
-    private final UserService userService;
-    private final EventService eventService;
+    private final UserRepository userRepository;
+    private final EventRepository eventRepository;
 
     @Override
+    @Transactional
     public CommentDto postComment(CommentDto commentDto, int userId, int eventId) {
         User user = getUserOrThrow(userId);
         Event event = getEventOrThrow(eventId);
         Comment comment = CommentMapper.dtoToComment(commentDto, user, event);
         commentRepository.save(comment);
         return CommentMapper.commentToDto(comment);
-    }
-
-    @Override
-    public List<CommentDto> getCommentsForEvent(int eventId) {
-        getEventOrThrow(eventId);
-        List<Comment> comments = commentRepository.findAllByEvent_Id(eventId);
-        return CommentMapper.commentToDtoList(comments);
     }
 
     @Override
@@ -47,9 +44,13 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    @Transactional
     public void deleteComment(int commentId, int eventId) {
-        getEventOrThrow(eventId);
-        getCommentOrThrow(commentId);
+        Event event = getEventOrThrow(eventId);
+        Comment comment = getCommentOrThrow(commentId);
+        if (!event.getComments().contains(comment)) {
+            throw new ConditionsNotMetException("Этот комментарий не принадлежит к этому событию");
+        }
         commentRepository.deleteCommentByEvent_IdAndId(eventId, commentId);
     }
 
@@ -57,14 +58,16 @@ public class CommentServiceImpl implements CommentService {
      * Проверка наличия пользователя в базе
      */
     private User getUserOrThrow(int userId) {
-        return userService.getUserOrThrow(userId);
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new EntryNotFoundException("Отсутствует пользователь с id: " + userId));
     }
 
     /**
      * Проверка наличия события в базе
      */
-    private Event getEventOrThrow(int eventId) {
-        return eventService.getEventOrThrow(eventId);
+    public Event getEventOrThrow(int eventId) {
+        return eventRepository.findById(eventId).orElseThrow(() ->
+                new EntryNotFoundException("Отсутствует событие с id: " + eventId));
     }
 
     private Comment getCommentOrThrow(int commentId) {
